Easy-Ocr/requirements.txt
pillow
easyocr
streamlit
